requirements.txt
pandas>=2.0.0
requests>=2.31.0
numpy>=1.24.0
google-genai>=0.4.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0
ddgs>=9.14.0
